1. Directory creation for pdf-only and data collection is added 2. xml compare is added with more logs and status sending report if its failed for any stage and with any exception. 3. added date and time in terminal messgae view.

+++ src/Rcvr_AAMQ/Utils.java
@@ -426,11 +426,11 @@ public class Utils {
 	   
 	   
 	   
-	   public boolean CreateNewDirectory(String newFolderPath) throws IOException
+	   public boolean CreateNewDirectory(String newFolderPath, boolean deleteExistDirectory) throws IOException
 	   {
 		   boolean success;
 		   File dirPath = new File(newFolderPath);
-		   if(dirPath.exists())
+		   if(dirPath.exists() && deleteExistDirectory)
 			   DeleteDirectory(newFolderPath);
 		   success = (new File(newFolderPath)).mkdirs();
 		   return success;
